Improved users.services.ts

+++ server/src/services/users.services.ts
@@ -1,16 +1,19 @@
-import {type User, UserWithoutId, type PartialUser} from '../common/users.interface.js';
+import {type User, type PartialUser} from '../common/users.interface.js';
 import * as Dao from '../DAO/users.dao.js';
 import * as mapping from '../mapping/users.mapping.js';
 import HttpError from '../common/error.class.js';
+import {type FindManyOptions} from 'typeorm';
+import type UserEntity from '../entities/users.entity';
 
 export async function get(
     id?: number,
-    filter?: {status?: string},
+    filter?: FindManyOptions<UserEntity>,
     page?: number,
     limit?: number,
 ) {
     if (id) {
         const user = await Dao.findUserById(id);
+
         if (!user) {
             throw new HttpError(`User with id ${id} not found`, 404);
         }
@@ -19,22 +22,26 @@ export async function get(
     }
 
     const users = await Dao.findAllUsers(filter, page, limit);
+
     return users.map(user => mapping.mapUserEntityToUser(user));
 }
 
 export async function create(userData: User) {
-    const user = await Dao.createUser(mapping.mapUserToUserEntity(userData));
-    return mapping.mapUserEntityToUser(user);
+    const userEntity = mapping.mapUserToUserEntity(userData);
+    const createdUser = await Dao.createUser(userEntity);
+
+    return mapping.mapUserEntityToUser(createdUser);
 }
 
 export async function update(id: number, userData: PartialUser) {
     const oldUserEntity = await Dao.findUserById(id);
+
     if (!oldUserEntity) {
         throw new HttpError(`User with id ${id} not found`, 404);
     }
 
     const oldUser = mapping.mapUserEntityToUser(oldUserEntity);
-    const updatedUser = await Dao.updateUser(mapping.mapUserToUserEntity({
+    const updatedUserData: User = {
         id,
         username: userData.username ? userData.username : oldUser.username,
         firstName: userData.firstName ? userData.firstName : oldUser.firstName,
@@ -42,11 +49,22 @@ export async function update(id: number, userData: PartialUser) {
         email: userData.email ? userData.email : oldUser.email,
         password: userData.password ? userData.password : oldUser.password,
         token: userData.token ? userData.token : oldUser.token,
-    }));
+    };
+
+    const updatedUserEntity = mapping.mapUserToUserEntity(updatedUserData);
+    const updatedUser = await Dao.updateUser(updatedUserEntity);
+
     return mapping.mapUserEntityToUser(updatedUser);
 }
 
 export async function remove(id: number) {
+    const existedUser = await Dao.findUserById(id);
+
+    if (!existedUser) {
+        throw new HttpError(`User with id ${id} not found`, 404);
+    }
+
     await Dao.deleteUser(id);
-    return 'user deleted';
+
+    return 'User deleted';
 }
